Move file-cluster-service config options to ConfigPhase.RUN_TIME
Fixes #8391
